refactor(splitter): align file organization and responsibilities with guidelines
Extract the Root sizes state machine into hooks/use-splitter-state.ts so
SplitterRoot is a thin provider; pull the pure deriveInitialSizes and
pickCollapseTarget helpers into utils/ with sibling specs (matching the
existing clampedResize treatment). Add components/, hooks/, and utils/
barrels and route imports through them. Move the context module to the
component root (.tsx->.ts) and relocate SplitterContextValue into the types
file. De-duplicate the slot prop types. Add @supportsStyleProps and per-part
JSDoc, and fix the __layoutRef typo in the useSplitterLayout example.

No behavior change: typecheck, lint, and 37 unit tests (+14 new) pass.
