+++ packages/nimbus/src/components/splitter/components/index.ts
@@ -1,0 +1,3 @@
+export { SplitterRoot } from "./splitter.root";
+export { SplitterPane } from "./splitter.pane";
+export { SplitterHandle } from "./splitter.handle";
