+++ packages/nimbus/src/components/splitter/hooks/index.ts
@@ -1,0 +1,7 @@
+export { useSplitterLayout } from "./use-splitter-layout";
+export type {
+  UseSplitterLayoutOptions,
+  UseSplitterLayoutResult,
+} from "./use-splitter-layout";
+export { useSplitterContext } from "./use-splitter-context";
+export { useSplitterState } from "./use-splitter-state";
