+++ packages/nimbus/src/components/splitter/index.ts
@@ -1,7 +1,7 @@
 export * from "./splitter";
 export * from "./splitter.types";
-export { useSplitterLayout } from "./hooks/use-splitter-layout";
+export { useSplitterLayout } from "./hooks";
 export type {
   UseSplitterLayoutOptions,
   UseSplitterLayoutResult,
-} from "./hooks/use-splitter-layout";
+} from "./hooks";
