+++ packages/nimbus/src/components/splitter/utils/index.ts
@@ -1,0 +1,4 @@
+export { clampedResize } from "./clamped-resize";
+export type { ClampedResizeArgs, ClampedResizeOptions } from "./clamped-resize";
+export { deriveInitialSizes } from "./derive-initial-sizes";
+export { pickCollapseTarget } from "./pick-collapse-target";
